[MM-40144] Restructure tests, add test case numbers, some cleanup
