+++ .eslintrc.json
@@ -58,12 +58,8 @@
         {
             "files": [
                 "webpack.config.renderer.js",
-                "e2e/specs/spellchecker_test.js",
-                "e2e/specs/app_test.js",
-                "e2e/specs/security_test.js",
-                "e2e/specs/permisson_test.js",
-                "e2e/specs/browser/index_test.js",
-                "e2e/specs/browser/settings_test.js",
+                "e2e/specs/startup/app.test.js",
+                "e2e/specs/settings.test.js",
                 "e2e/modules/utils.js",
                 "e2e/modules/environment.js",
                 "CHANGELOG.md",
