--- e2e/specs/index.js
@@ -1,13 +1,0 @@
-// Copyright (c) 2016-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-
-import './app_test.js';
-import './config_test.js';
-import './deeplink_test.js';
-import './mattermost_test.js';
-import './menu_test.js';
-import './window_test.js';
-
-import './browser/index_test.js';
-import './browser/modal_test.js';
-import './browser/settings_test.js';
