+++ e2e/specs/startup/app.test.js
@@ -4,9 +4,9 @@
 
 'use strict';
 
-const env = require('../modules/environment');
+const env = require('../../modules/environment');
 
-describe('application', function desc() {
+describe('startup/app', function desc() {
     this.timeout(30000);
 
     beforeEach(async () => {
@@ -21,19 +21,19 @@ describe('application', function desc() {
         }
     });
 
-    it('should show the new server modal when no servers exist', async () => {
+    it('MM-T4399_1 should show the new server modal when no servers exist', async () => {
         const newServerModal = this.app.windows().find((window) => window.url().includes('newServer'));
         const modalTitle = await newServerModal.innerText('#newServerModal .modal-title');
         modalTitle.should.equal('Add Server');
     });
 
-    it('should show no servers configured in dropdown when no servers exist', async () => {
+    it('MM-T4399_2 should show no servers configured in dropdown when no servers exist', async () => {
         const mainWindow = this.app.windows().find((window) => window.url().includes('index'));
         const dropdownButtonText = await mainWindow.innerText('.TeamDropdownButton');
         dropdownButtonText.should.equal('No servers configured');
     });
 
-    it('should be stopped when the app instance already exists', (done) => {
+    it('MM-T4400 should be stopped when the app instance already exists', (done) => {
         const secondApp = env.getApp();
 
         // In the correct case, 'start().then' is not called.
