+++ webpack.config.test.js
@@ -16,8 +16,7 @@ const WEBSERVER_PORT = 9001;
 
 module.exports = merge(base, {
     entry: {
-        test: glob.sync('./src/**/*.test.js'),
-        e2e: './e2e/specs/index.js',
+        e2e: glob.sync('./e2e/specs/**/*.test.js'),
     },
     output: {
         path: path.resolve(__dirname, 'dist/tests'),
